Move linalg submodule to msm module.
