+++ src/msmhelper/msm/utils/__init__.py
@@ -1,0 +1,2 @@
+# -*- coding: utf-8 -*-
+"""  --8<-- "README.md" """
