+++ src/msmhelper/msm/utils/linalg.py
